requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4,<2
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic>=1.10,<2
email-validator
python-multipart
argon2-cffi
